Add files via upload
updated the scraper with a self balancing weights to return only the best unused links. if it finds too little it lowers the weights, if it finds too many, it raises the weights. used urls are now stored in a locally stored sql file, and cleaned every 3 months. the csv now stores a unique signature, the title, the content ( already cleaned up for tts, run through replacement to remove slang, and then again through dict for stripping strange characters) and a combined tts with prompts on what to save the mp3 as. For refinement I will be replacing the last field with a confidence score so you can sort by engagement, Deliberately keeping it csv to keep it human readable at a basic entry level.